Consistently handle top-level arrays
