+++ src/main/php/text/json/DenseFormat.class.php
@@ -18,44 +18,4 @@ class DenseFormat extends Format {
   public function __construct($options= 0) {
     parent::__construct(',', ':', $options);
   }
-
-  /**
-   * Formats an array
-   *
-   * @param  var[] $value
-   * @return string
-   */
-  protected function formatArray($value) {
-    $r= '[';
-    $next= false;
-    foreach ($value as $element) {
-      if ($next) {
-        $r.= ',';
-      } else {
-        $next= true;
-      }
-      $r.= $this->representationOf($element);
-    }
-    return $r.']';
-  }
-
-  /**
-   * Formats an object
-   *
-   * @param  [:var] $value
-   * @return string
-   */
-  protected function formatObject($value) {
-    $r= '{';
-    $next= false;
-    foreach ($value as $key => $mapped) {
-      if ($next) {
-        $r.= ',';
-      } else {
-        $next= true;
-      }
-      $r.= $this->representationOf($key).':'.$this->representationOf($mapped);
-    }
-    return $r.'}';
-  }
 }
