+++ src/main/php/text/json/Format.class.php
@@ -1,11 +1,12 @@
 <?php namespace text\json;
 
+use StdClass;
 use lang\{IllegalArgumentException, Value};
 
 /**
  * JSON format
  *
- * @test  xp://text.json.unittest.FormatFactoryTest
+ * @test  text.json.unittest.FormatFactoryTest
  */
 abstract class Format implements Value {
   const ESCAPE_SLASHES = -65;  // ~JSON_UNESCAPED_SLASHES
@@ -54,20 +55,24 @@ abstract class Format implements Value {
   }
 
   /**
-   * Formats an array
-   *
-   * @param  var[] $value
-   * @return string
-   */
-  protected abstract function formatArray($value);
-
-  /**
    * Formats an object
    *
    * @param  [:var] $value
    * @return string
    */
-  protected abstract function formatObject($value);
+  protected function formatObject($value) {
+    $r= $this->open('{');
+    $next= false;
+    foreach ($value as $key => $mapped) {
+      if ($next) {
+        $r.= $this->comma;
+      } else {
+        $next= true;
+      }
+      $r.= $this->representationOf($key).$this->colon.$this->representationOf($mapped);
+    }
+    return $r.$this->close('}');
+  }
 
   /**
    * Open an array or object
@@ -108,7 +113,17 @@ abstract class Format implements Value {
       if (empty($value)) {
         return '[]';
       } else if (0 === key($value)) {
-        return $this->formatArray($value);
+        $r= $this->open('[');
+        $next= false;
+        foreach ($value as $element) {
+          if ($next) {
+            $r.= $this->comma;
+          } else {
+            $next= true;
+          }
+          $r.= $this->representationOf($element);
+        }
+        return $r.$this->close(']');
       } else {
         return $this->formatObject($value);
       }
@@ -118,7 +133,7 @@ abstract class Format implements Value {
       return 'true';
     } else if (false === $value) {
       return 'false';
-    } else if ($value instanceof \stdclass) {
+    } else if ($value instanceof StdClass) {
       $cast= (array)$value;
       if (empty($cast)) {
         return '{}';
