+++ src/main/php/text/json/WrappedFormat.class.php
@@ -24,55 +24,11 @@ class WrappedFormat extends Format {
   }
 
   /**
-   * Formats an array
+   * Open an array or object
    *
-   * @param  var[] $value
-   * @return string
+   * @param  string $token either `[` or `{`
+   * @param  string
    */
-  protected function formatArray($value) {
-    $comma= $this->comma;
-    $this->comma= ', ';
-    $r= '[';
-    $next= false;
-    foreach ($value as $element) {
-      if ($next) {
-        $r.= $this->comma;
-      } else {
-        $next= true;
-      }
-      $r.= $this->representationOf($element);
-    }
-    $this->comma= $comma;
-    return $r.']';
-  }
-
-  /**
-   * Formats an object
-   *
-   * @param  [:var] $value
-   * @return string
-   */
-  protected function formatObject($value) {
-    $comma= $this->comma;
-    $indent= str_repeat($this->indent, $this->level);
-    $this->comma= ",\n".$indent;
-    $r= "{\n".$indent;
-    $next= false;
-    $this->level++;
-    foreach ($value as $key => $mapped) {
-      if ($next) {
-        $r.= $this->comma;
-      } else {
-        $next= true;
-      }
-      $r.= $this->representationOf($key).$this->colon.$this->representationOf($mapped);
-    }
-    $this->level--;
-    $indent= str_repeat($this->indent, $this->level - 1);
-    $this->comma= $comma;
-    return $r."\n".$indent.'}';
-  }
-
   public function open($token) {
     if ('{' === $token || empty($this->stack)) {
       $this->stack[]= $this->comma;
@@ -86,6 +42,12 @@ class WrappedFormat extends Format {
     }
   }
 
+  /**
+   * Close an array or object
+   *
+   * @param  string $token either `]` or `}`
+   * @param  string
+   */
   public function close($token) {
     $this->comma= array_pop($this->stack);
     if ('}' === $token || empty($this->stack)) {
